add test-utilities, initial draft for build and serve commands
