+++ bin/scripts/build.js
@@ -1,8 +1,3 @@
-// #!/usr/bin/env node
-// console.log('Toolkit - Creating Build');
-// // process.env.BABEL_DISABLE_CACHE = 0;
-// require('babel-register');
-// global.toolkitCli = {
-//   isDev: false,
-// };
-// require('../dist/build');
+#!/usr/bin/env node
+require('../utils/bootstrap');
+require('../../dist/build');
