+++ package.json
@@ -83,6 +83,7 @@
     "handlebars-loader": "^1.3.0",
     "html-webpack-plugin": "^2.21.0",
     "image-webpack-loader": "^1.8.0",
+    "jsdom": "^9.4.2",
     "json-loader": "^0.5.4",
     "jsx-control-statements": "^3.1.2",
     "manifest-revision-webpack-plugin": "^0.3.0",
